refactor(registry): extract matchUserSkillsForTask to shared skill-matcher (#bug23.0)
Move the keyword-based skill matcher (#bug22.3) out of `hooks-tools.ts`
and into `src/registry/skill-matcher.ts` so the upcoming `swarm_init`
auto-agent-selection (#bug23) can reuse the exact same scorer instead
of duplicating the logic.

Function semantics are unchanged — the extracted module exports
`matchUserSkillsForTask`, `tokenizeForSkillMatching`,
`buildSkillKeywordBag`, `COMMON_STOPWORDS`, and the `UserSkillMatch`
type. `hooks-tools.ts` now imports these from the shared module; all
existing #bug22.3 regression tests continue to pass without
modification.

Why extract instead of duplicate: `hooks_route` and `swarm_init` both
need to recommend user-installed content for the same task. If the
two tools used independent copies of the matcher, a future tweak to
the scoring (stopwords, weight, threshold) in one place would silently
diverge from the other and the caller would see contradictory advice.
One source of truth keeps them honest.
