+++ v3/@claude-flow/cli/src/mcp-tools/hooks-tools.ts
@@ -8,7 +8,8 @@ import { dirname, join, resolve } from 'path';
 import { homedir } from 'os';
 import { type MCPTool, getProjectCwd } from './types.js';
 import { validateIdentifier, validateText, validatePath } from './validate-input.js';
-import { scanClaudeCodeRegistry, type UserSkill, type UserAgent } from '../registry/claude-code-registry.js';
+import { scanClaudeCodeRegistry } from '../registry/claude-code-registry.js';
+import { matchUserSkillsForTask, type UserSkillMatch } from '../registry/skill-matcher.js';
 
 // Real vector search functions - lazy loaded to avoid circular imports
 let searchEntriesFn: ((options: {
@@ -885,132 +886,11 @@ export const hooksPostCommand: MCPTool = {
 };
 
 /**
- * #bug22.3 — Match user-installed skills/agents against a task description.
- *
- * MVP keyword approach (intentionally simple — see TODO below). For each
- * user skill we tokenize:
- *   - The skill `name` (e.g. "polymarket-analyzer" → ["polymarket",
- *     "analyzer"]).
- *   - The first ~20 words of `description` (when present), filtered to
- *     informative tokens (>= 4 chars, alphanumerics, deduped).
- *
- * Each task token that overlaps the skill's bag scores +1, with the bag
- * built so a direct hit on a unique-noun in the skill name is weighted
- * heavily (the skill name itself contributes 2× — it's the most reliable
- * signal that the user wrote a skill specifically for this).
- *
- * Returns the top-K matches sorted by score (>= 1). Empty array means no
- * user skill should be considered.
- *
- * TODO(bug22-followup): replace this with proper embedding-based scoring
- * once we route through `bridgeStorePattern` for skill descriptions. The
- * keyword scorer is a bridge — it covers the obvious "polymarket" /
- * "kali" / "geo" / "ceo" cases that the hardcoded built-in catalog
- * completely misses, but won't catch synonyms or paraphrases.
+ * #bug23.0 — `matchUserSkillsForTask` was extracted to
+ * `../registry/skill-matcher.ts` so `swarm_init` (#bug23) can reuse the
+ * same scorer. The function semantics are unchanged from #bug22.3 — see
+ * the extracted module for full docs.
  */
-interface UserSkillMatch {
-  type: 'skill' | 'agent';
-  name: string;
-  score: number;
-  description?: string;
-  matchedKeywords: string[];
-}
-
-function tokenizeForSkillMatching(text: string): string[] {
-  return text
-    .toLowerCase()
-    // Split on non-alphanumerics so dashes / underscores yield separate tokens.
-    .split(/[^a-z0-9]+/)
-    .filter((t) => t.length >= 3);
-}
-
-const COMMON_STOPWORDS = new Set([
-  'the', 'and', 'for', 'with', 'that', 'this', 'use', 'when', 'how', 'has',
-  'are', 'you', 'can', 'all', 'any', 'should', 'will', 'have', 'been', 'was',
-  'were', 'into', 'from', 'about', 'task', 'tasks', 'agent', 'agents', 'skill',
-  'skills', 'using', 'used', 'used', 'than', 'then', 'their', 'there', 'where',
-]);
-
-function buildSkillKeywordBag(skill: UserSkill | UserAgent): { tokens: Set<string>; nameTokens: Set<string> } {
-  const nameTokens = new Set(tokenizeForSkillMatching(skill.name));
-  const bag = new Set<string>(nameTokens);
-
-  // Pull tokens from description but cap how much we mine — long
-  // descriptions otherwise drown out the skill-name signal.
-  if (skill.description) {
-    const descTokens = tokenizeForSkillMatching(skill.description).slice(0, 40);
-    for (const t of descTokens) {
-      if (!COMMON_STOPWORDS.has(t)) bag.add(t);
-    }
-  }
-  return { tokens: bag, nameTokens };
-}
-
-function matchUserSkillsForTask(
-  task: string,
-  context: string | undefined,
-  skills: UserSkill[],
-  agents: UserAgent[],
-  topK = 5,
-): UserSkillMatch[] {
-  const taskTokens = new Set([
-    ...tokenizeForSkillMatching(task),
-    ...(context ? tokenizeForSkillMatching(context) : []),
-  ]);
-  if (taskTokens.size === 0) return [];
-
-  const candidates: UserSkillMatch[] = [];
-
-  for (const skill of skills) {
-    const { tokens, nameTokens } = buildSkillKeywordBag(skill);
-    let score = 0;
-    const matched: string[] = [];
-    for (const t of taskTokens) {
-      if (tokens.has(t)) {
-        // Name-token match weighted 2× because it's the strongest signal.
-        const weight = nameTokens.has(t) ? 2 : 1;
-        score += weight;
-        matched.push(t);
-      }
-    }
-    if (score >= 1) {
-      candidates.push({
-        type: 'skill',
-        name: skill.name,
-        score,
-        description: skill.description,
-        matchedKeywords: matched,
-      });
-    }
-  }
-
-  // Also consider user agents — many agents (polymarket-analyzer, ceo,
-  // polybot-ops) are addressable directly even if there's no skill for them.
-  for (const agent of agents) {
-    const { tokens, nameTokens } = buildSkillKeywordBag(agent);
-    let score = 0;
-    const matched: string[] = [];
-    for (const t of taskTokens) {
-      if (tokens.has(t)) {
-        const weight = nameTokens.has(t) ? 2 : 1;
-        score += weight;
-        matched.push(t);
-      }
-    }
-    if (score >= 1) {
-      candidates.push({
-        type: 'agent',
-        name: agent.name,
-        score,
-        description: agent.description,
-        matchedKeywords: matched,
-      });
-    }
-  }
-
-  candidates.sort((a, b) => b.score - a.score);
-  return candidates.slice(0, topK);
-}
 
 export const hooksRoute: MCPTool = {
   name: 'hooks_route',
